Move ScalabilityMode and RtcpFeedback in capability API module
