+++ crates/native/src/api/mod.rs
@@ -15,7 +15,8 @@ use flutter_rust_bridge::for_generated::FLUTTER_RUST_BRIDGE_RUNTIME_VERSION;
 use libwebrtc_sys as sys;
 
 pub use self::capability::{
-    RtpCapabilities, RtpCodecCapability, RtpHeaderExtensionCapability,
+    RtcpFeedback, RtcpFeedbackMessageType, RtcpFeedbackType, RtpCapabilities,
+    RtpCodecCapability, RtpHeaderExtensionCapability, ScalabilityMode,
     get_rtp_receiver_capabilities, get_rtp_sender_capabilities,
     set_codec_preferences,
 };
@@ -543,306 +544,6 @@ impl From<sys::IceCandidateStats> for IceCandidateStats {
             priority,
             url,
             relay_protocol: None,
-        }
-    }
-}
-
-impl From<sys::RtcpFeedbackType> for RtcpFeedbackType {
-    fn from(value: sys::RtcpFeedbackType) -> Self {
-        match value {
-            sys::RtcpFeedbackType::CCM => Self::Ccm,
-            sys::RtcpFeedbackType::LNTF => Self::Lntf,
-            sys::RtcpFeedbackType::NACK => Self::Nack,
-            sys::RtcpFeedbackType::REMB => Self::Remb,
-            sys::RtcpFeedbackType::TRANSPORT_CC => Self::TransportCC,
-            _ => unreachable!(),
-        }
-    }
-}
-impl From<sys::RtcpFeedbackMessageType> for RtcpFeedbackMessageType {
-    fn from(value: sys::RtcpFeedbackMessageType) -> Self {
-        match value {
-            sys::RtcpFeedbackMessageType::GENERIC_NACK => Self::GenericNACK,
-            sys::RtcpFeedbackMessageType::PLI => Self::Pli,
-            sys::RtcpFeedbackMessageType::FIR => Self::Fir,
-            _ => unreachable!(),
-        }
-    }
-}
-
-/// [RTCP] feedback message intended to enable congestion control for
-/// interactive real-time traffic using [RTP].
-///
-/// [RTCP]: https://en.wikipedia.org/wiki/RTP_Control_Protocol
-/// [RTP]: https://en.wikipedia.org/wiki/Real-time_Transport_Protocol
-#[derive(Debug)]
-pub struct RtcpFeedback {
-    /// Message type of this [`RtcpFeedback`].
-    pub message_type: Option<RtcpFeedbackMessageType>,
-
-    /// Kind of this [`RtcpFeedback`].
-    pub kind: RtcpFeedbackType,
-}
-
-impl From<sys::RtcpFeedback> for RtcpFeedback {
-    fn from(value: sys::RtcpFeedback) -> Self {
-        Self {
-            message_type: value.message_type().map(Into::into),
-            kind: value.kind().into(),
-        }
-    }
-}
-
-/// [ScalabilityMode][0] representation.
-///
-/// [0]: https://tinyurl.com/35ae3mbe
-#[derive(Debug, Eq, Hash, PartialEq)]
-#[repr(u8)]
-pub enum ScalabilityMode {
-    /// [ScalabilityMode.L1T1][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L1T1*
-    L1T1 = 0,
-
-    /// [ScalabilityMode.L1T2][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L1T2*
-    L1T2,
-
-    /// [ScalabilityMode.L1T3][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L1T3*
-    L1T3,
-
-    /// [ScalabilityMode.L2T1][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T1*
-    L2T1,
-
-    /// [ScalabilityMode.L2T1h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T1*
-    L2T1h,
-
-    /// [ScalabilityMode.L2T1_KEY][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T1_KEY*
-    L2t1Key,
-
-    /// [ScalabilityMode.L2T2][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T2h*
-    L2T2,
-
-    /// [ScalabilityMode.L2T2h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T2*
-    L2T2h,
-
-    /// [ScalabilityMode.L2T2_KEY][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T2_KEY*
-    L2T2Key,
-
-    /// [ScalabilityMode.L2T2_KEY_SHIFT][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T2_KEY_SHIFT*
-    L2T2KeyShift,
-
-    /// [ScalabilityMode.L2T3][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T3*
-    L2T3,
-
-    /// [ScalabilityMode.L2T3h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T3*
-    L2T3h,
-
-    /// [ScalabilityMode.L2T3_KEY][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L2T3_KEY*
-    L2T3Key,
-
-    /// [ScalabilityMode.L3T1][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L3T1*
-    L3T1,
-
-    /// [ScalabilityMode.L3T1h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L3T1*
-    L3T1h,
-
-    /// [ScalabilityMode.L3T1_KEY][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L3T1_KEY*
-    L3T1Key,
-
-    /// [ScalabilityMode.L3T2][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L3T2h*
-    L3T2,
-
-    /// [ScalabilityMode.L3T2h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L3T2*
-    L3T2h,
-
-    /// [ScalabilityMode.L3T2_KEY][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L3T2_KEY*
-    L3T2Key,
-
-    /// [ScalabilityMode.kL3T3][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#kL3T3*
-    L3T3,
-
-    /// [ScalabilityMode.kL3T3h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#kL3T3*
-    L3T3h,
-
-    /// [ScalabilityMode.kL3T3_KEY][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#L3T3_KEY*
-    L3T3Key,
-
-    /// [ScalabilityMode.kS2T1][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#kS2T1*
-    S2T1,
-
-    /// [ScalabilityMode.kS2T1h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#kS2T1*
-    S2T1h,
-
-    /// [ScalabilityMode.kS2T2][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#kS2T2*
-    S2T2,
-
-    /// [ScalabilityMode.kS2T2h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#kS2T2*
-    S2T2h,
-
-    /// [ScalabilityMode.S2T3][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#S2T3h*
-    S2T3,
-
-    /// [ScalabilityMode.S2T3h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#S2T3*
-    S2T3h,
-
-    /// [ScalabilityMode.S3T1h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#S3T1*
-    S3T1,
-
-    /// [ScalabilityMode.S3T1h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#S3T1*
-    S3T1h,
-
-    /// [ScalabilityMode.S3T2][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#S3T2*
-    S3T2,
-
-    /// [ScalabilityMode.S3T2h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#S3T2*
-    S3T2h,
-
-    /// [ScalabilityMode.S3T3][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#S3T3*
-    S3T3,
-
-    /// [ScalabilityMode.S3T3h][0] mode.
-    ///
-    /// [0]: https://w3.org/TR/webrtc-svc#S3T3*
-    S3T3h,
-}
-
-/// Possible types of an [`RtcpFeedback`].
-#[derive(Debug, Eq, Hash, PartialEq)]
-#[repr(i32)]
-pub enum RtcpFeedbackType {
-    /// Codec control messages.
-    Ccm,
-
-    /// Loss notification feedback.
-    Lntf,
-
-    /// Negative acknowledgemen.
-    Nack,
-
-    /// Receiver estimated maximum bitrate.
-    Remb,
-
-    /// Transport wide congestion control.
-    TransportCC,
-}
-
-/// Possible message types of an [`RtcpFeedback`], when is type is
-/// [`RtcpFeedbackType::Nack`] or [`RtcpFeedbackType::Ccm`].
-#[derive(Debug, Eq, Hash, PartialEq)]
-#[repr(i32)]
-pub enum RtcpFeedbackMessageType {
-    /// Equivalent to `{ type: "nack", parameter: undefined }` in ORTC.
-    GenericNACK,
-
-    /// Usable with [`RtcpFeedbackType::Nack`].
-    Pli,
-
-    /// Usable with [`RtcpFeedbackType::Ccm`].
-    Fir,
-}
-
-impl From<sys::ScalabilityMode> for ScalabilityMode {
-    fn from(value: sys::ScalabilityMode) -> Self {
-        match value {
-            sys::ScalabilityMode::kL1T1 => Self::L1T1,
-            sys::ScalabilityMode::kL1T2 => Self::L1T2,
-            sys::ScalabilityMode::kL1T3 => Self::L1T3,
-            sys::ScalabilityMode::kL2T1 => Self::L2T1,
-            sys::ScalabilityMode::kL2T1h => Self::L2T1h,
-            sys::ScalabilityMode::kL2T1_KEY => Self::L2t1Key,
-            sys::ScalabilityMode::kL2T2 => Self::L2T2,
-            sys::ScalabilityMode::kL2T2h => Self::L2T2h,
-            sys::ScalabilityMode::kL2T2_KEY => Self::L2T2Key,
-            sys::ScalabilityMode::kL2T2_KEY_SHIFT => Self::L2T2KeyShift,
-            sys::ScalabilityMode::kL2T3 => Self::L2T3,
-            sys::ScalabilityMode::kL2T3h => Self::L2T3h,
-            sys::ScalabilityMode::kL2T3_KEY => Self::L2T3Key,
-            sys::ScalabilityMode::kL3T1 => Self::L3T1,
-            sys::ScalabilityMode::kL3T1h => Self::L3T1h,
-            sys::ScalabilityMode::kL3T1_KEY => Self::L3T1Key,
-            sys::ScalabilityMode::kL3T2 => Self::L3T2,
-            sys::ScalabilityMode::kL3T2h => Self::L3T2h,
-            sys::ScalabilityMode::kL3T2_KEY => Self::L3T2Key,
-            sys::ScalabilityMode::kL3T3 => Self::L3T3,
-            sys::ScalabilityMode::kL3T3h => Self::L3T3h,
-            sys::ScalabilityMode::kL3T3_KEY => Self::L3T3Key,
-            sys::ScalabilityMode::kS2T1 => Self::S2T1,
-            sys::ScalabilityMode::kS2T1h => Self::S2T1h,
-            sys::ScalabilityMode::kS2T2 => Self::S2T2,
-            sys::ScalabilityMode::kS2T2h => Self::S2T2h,
-            sys::ScalabilityMode::kS2T3 => Self::S2T3,
-            sys::ScalabilityMode::kS2T3h => Self::S2T3h,
-            sys::ScalabilityMode::kS3T1 => Self::S3T1,
-            sys::ScalabilityMode::kS3T1h => Self::S3T1h,
-            sys::ScalabilityMode::kS3T2 => Self::S3T2,
-            sys::ScalabilityMode::kS3T2h => Self::S3T2h,
-            sys::ScalabilityMode::kS3T3 => Self::S3T3,
-            sys::ScalabilityMode::kS3T3h => Self::S3T3h,
-            _ => unreachable!(),
         }
     }
 }
